Update Plugins for Near with VATZ Plugins
